RAG_ATIVIDADE_FIAP/requirements.txt
azure-functions
azure-search-documents==11.6.0b4
pypdf
python-dotenv
langchain
langchain-openai
pydantic
